requirements.txt
django==1.4
django-compressor==1.2
django-dajax==0.9.2
django-dajaxice==0.5.3

discogs-client==1.1.1
PIL==1.1.7
BeautifulSoup<=4.0
requests==0.14.2

